update session role

+++ admin-kurir/hapus_kurir.php
@@ -1,6 +1,13 @@
 <?php
 include 'config/koneksi.php';
 
+session_start();
+if ($_SESSION['level'] !== 'admin') {
+    echo "Akses ditolak. Halaman ini hanya untuk admin.";
+    exit;
+}
+
+
 $id = $_GET['id'];
 
 $query = "DELETE FROM tbl_data_kurir WHERE id_kurir = '$id'";
